1.9) Simulating memory
Full simulation of segmented memory.
The x86 binary is now loaded into the code segment (0x0) at startup.
The stack segment is located at 0x10000 (physical).
The data segment is located at 0x20000 (physical).
The extra segment is initialized to point to the data segment.

As there was no challenge homework for segmented memory, it has not been thouroughly tested. Unless an explicit segment is picked, all memory operations currently default to the Data segment unless BP is used as a base register, in which case the stack segment is used.

String memory references are not implemented. There are currently no stack operations implemented.